B41-F24-PA34

+++ backend/src/main/groovy/com/backend/gbp/domain/billing/JobItems.groovy
@@ -58,7 +58,7 @@ class JobItems extends AbstractAuditingEntity implements Serializable {
 
 	@GraphQLQuery
 	@Column(name = "qty")
-	Integer qty
+	BigDecimal qty
 
 	@GraphQLQuery
 	@Column(name = "cost")
