+++ backend/src/main/groovy/com/backend/gbp/domain/services/ServiceItems.groovy
@@ -39,7 +39,7 @@ class ServiceItems extends AbstractAuditingEntity implements Serializable {
 
 	@GraphQLQuery
 	@Column(name = "qty")
-	Integer qty
+	BigDecimal qty
 
 	@GraphQLQuery
 	@Column(name = "wcost")
